Restructure Logger to new paradigm
