+++ src/Core/Logger/Type/AbstractLogger.php
@@ -19,7 +19,7 @@
  *
  */
 
-namespace Friendica\Util\Logger;
+namespace Friendica\Core\Logger\Type;
 
 use Friendica\Util\Introspection;
 use Friendica\Util\Strings;
