+++ src/Core/Logger/Type/Monolog/IntrospectionProcessor.php
@@ -19,7 +19,7 @@
  *
  */
 
-namespace Friendica\Util\Logger\Monolog;
+namespace Friendica\Core\Logger\Type\Monolog;
 
 use Friendica\Util\Introspection;
 use Monolog\Logger;
