+++ src/Core/Logger/Type/StreamLogger.php
@@ -19,7 +19,7 @@
  *
  */
 
-namespace Friendica\Util\Logger;
+namespace Friendica\Core\Logger\Type;
 
 use Friendica\Util\DateTimeFormat;
 use Friendica\Util\FileSystem;
