+++ src/DI.php
@@ -243,7 +243,7 @@ abstract class DI
 	 */
 	public static function workerLogger()
 	{
-		return self::$dice->create(Util\Logger\WorkerLogger::class);
+		return self::$dice->create(Core\Logger\Type\WorkerLogger::class);
 	}
 
 	//
