+++ tests/src/Core/Logger/WorkerLoggerTest.php
@@ -19,10 +19,10 @@
  *
  */
 
-namespace Friendica\Test\src\Util\Logger;
+namespace Friendica\Test\src\Core\Logger;
 
+use Friendica\Core\Logger\Type\WorkerLogger;
 use Friendica\Test\MockedTest;
-use Friendica\Util\Logger\WorkerLogger;
 use Psr\Log\LoggerInterface;
 
 class WorkerLoggerTest extends MockedTest
